systemd: Add systemd units to automatically run at the right times
Ensuring the utility runs at all the times it needs to requires systemd
units. It has to run at boot and also after any resume from suspend. It
also has to run sufficiently late that the controller is up. As there's
no event I could find to listen for to know when the controller is up,
I had to use a time delay, driven by a systemd timer.

The units need to be edits to set the right device addresses at a
minimum.